Implemented router and added tailwind config
