+++ postcss.config.ts
@@ -1,0 +1,8 @@
+import type { Config } from 'postcss-load-config';
+
+export default {
+  plugins: {
+    '@tailwindcss/postcss': {},
+    autoprefixer: {},
+  }
+} satisfies Config;
